fix. can make transactions with a float amount

+++ public/queries/myQueries.php
@@ -101,7 +101,7 @@ class MyQueries {
         return $rows;
     }
 
-    public static function newTInsertQuery(PDO $conn, string $tipo, int $monto, string $concepto, int $id_user, int $id_account) : int
+    public static function newTInsertQuery(PDO $conn, string $tipo, float $monto, string $concepto, int $id_user, int $id_account) : int
     { 
         $sql = $conn->exec("INSERT INTO movimientos (tipo,monto,concepto,id_usuario,id_cuenta) VALUES ('$tipo','$monto','$concepto','$id_user','$id_account')");
         return $sql;
